counts number of flip of a bit

+++ 0x14-bit_manipulation/main.h
@@ -10,6 +10,6 @@ int get_bit(unsigned long int n, unsigned int index);
 int set_bit(unsigned long int *n, unsigned int index);
 int clear_bit(unsigned long int *n, unsigned int index);
 unsigned int flip_bits(unsigned long int n, unsigned long int m);
-int counter(int num);
+int counter(unsigned long int num);
 
 #endif
